requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
python-decouple==3.8
mixer==7.2.2
